Add flag for operations within eval call
pyexz3.py now appends code to the start of the file being tested to
redefine the `eval` method within the file's scope to set a variable in
`builtins`, which is then accessible later in the call stack. During
symbolic execution, if this variable is set when computing the result of
an operation involving a symbolic object, the engine now notes this and
outputs the variables and values of the objects.